Warn users not to modify parameter adminUsername and modify explanations about it
1. Add explanation of parameter adminUsername in ARM Template's parameters, and warn users not to modify it in line 132.
There is no explanation for the parameter of adminUsername previously. Actually, modifying it may result in failure in the installation of Elastic Runtime and will cause the Ops Manager VM inaccessible via SSH. (Even users try to reset SSH key or password with Microsoft Azure's Portal) 

2. Remove confused explanation about username while generate SSH RSA key pair in line 116.
`-C ubuntu` only specifies the `comment` of the key pair to be 'ubuntu' and has nothing to do with username.
Modify the 'username' `ubuntu` simultaneously in the comment of key pair and parameters for ARM Template won't help in problems mentioned in previous section.