Document ideal location for metadata backup files
Backing up these files within the same directories as the original will result in OpsMan failing to restart. We think this is because OpsManager attempts to load all yaml files in the respective directories. The instructions are using a different directory as example of where to take the backup, however it is not obvious/explicit in the docs.