Remove confusing syslog custom rule examples
We're likely to want to give the article a general once-over later, but this PR is to eliminate documentation for a couple of things we'd rather people not try to do with the custom rule field.

As we identify more specific scenarios useful to customers, we'll probably add trickier rule examples, but for now, we'd like to stick to changing which logs are forwarded.

We've already had an FE inquire about this feature based on these docs. The custom rule you'd need to actually change the egress format (not just write to the disk in a certain format) is actually fairly complicated to describe. And, we would prefer customers not get in the habit of manipulating the log format, especially since we'll be rearranging that config substantially soon, and rules that mess with egress format may not continue to work as expected.

Similarly, the rule to write certain logs to a local file is much more useful for development than for end users, and is something most PCF users shouldn't do.