Clarifying how to retrieve self-signed NSX CA Cert
The docs originally mentioned that you could provide the fully-qualified domain name for the NSX manager in this field if it was a self-signed certificate, but this doesn't work.  It wasn't clear how to retrieve this CA cert if you allowed NSX to generate its own certs during installation, so I've added an example command.