Update AWS Permissions Required by Terraform
This commit amends the permissions that are required for Terraform to
run the tasks that are defined in the officially supported Terraform
scripts. The required permissions have already been updated for other
projects, such as
[pcf-pipelines](https://github.com/pivotal-cf/terraforming-aws#aws-permissions),
but the official documentation is still using old naming conventions
for two permissions (permissions required for IAM and KMS operations).