Correct steps for NAT setup on GCP
Update the directions for setting up the NAT instances on GCP to conform
with the knowledge base article from June.  In June an issue was
discovered with the GCP NAT setup where in some cases (package updates),
the settings per directions will not persist. These directions will
ensure that settings persist and will be more resilient.

https://community.pivotal.io/s/article/Loss-of-network-connectivity-on-Google-Cloud-Platform-after-applying-security-patch-DSA-4208