Clarify use of AWS LBs
In several cases, the docs are assuming ELB always means a Classic ELB,
but our manual installation actually calls for an application load
balancer (ALB) for the web LB.

- instruct users to enter the target group name and not the load
balancer name in the resource config when using ALB and NLB
- Remove reference to port 4443, as ALBs support websockets and do not
require a separate load balancer on a separate port.  The default
loggregator port of 443 works with ALBs out of the box.
- Correct the name of the load balancer field in Operations Manager's
resource config
- Remove mention of supporting multiple load balancers (this doesn't
work)