Updated Prerequisites
Removed reference to the cookbook which is severely outdated (NSX 2.2) and is now causing problems in the field as it contains steps that no longer need to be performed, that create issues if they *are* performed.

Replaced with link to NSX-T/NCP/PAS support matrix site and link to VMware's NSX-T installation instructions.

Updated general wording of line to be non-version specific, since the versions mentioned were old.

Updated BOSH/Ops Manager line to remove version, since it was also old - PAS 2.5 mandates Ops Manager 2.5 for installation to be possible so do we need to specify (and update/forget to update) the version in this doc each time?

Same for mention of PAS version near the beginning.